cr30. remove_serverhack_macro. Remove _TEST from serverhack and cmocka.
