+++ cmocka/param_test.c
@@ -45,6 +45,30 @@ static void test_get_param_verbose_level2(void **state) {
 	assert_int_equal(2, get_param_verbose_level(args, 2));
 }
 
+static void test_get_param_mode_console_1(void **state) {
+	char *args[]= {"GrwolServerHack" };
+	int argc= 1;
+	(void) state; /* unused */
+
+	assert_int_equal(0,get_param_mode(args,argc,"--console"));
+}
+
+static void test_get_param_mode_console_2(void **state) {
+	char *args[]= {"GrwolServerHack", "--console" };
+	int argc= 2;
+	(void) state; /* unused */
+
+	assert_int_equal(1,get_param_mode(args,argc,"--console"));
+}
+
+static void test_get_param_mode_console_3(void **state) {
+	char *args[]= {"GrwolServerHack", "--console", "off" };
+	int argc= 3;
+	(void) state; /* unused */
+
+	assert_int_equal(0,get_param_mode(args,argc,"--console"));
+}
+
 int main(void)
 {
 	const struct CMUnitTest tests[] = {
@@ -52,6 +76,9 @@ int main(void)
 		cmocka_unit_test(test_get_param_verbose_level1),
 		cmocka_unit_test(test_get_param_verbose_level1a),
 		cmocka_unit_test(test_get_param_verbose_level2),
+		cmocka_unit_test(test_get_param_mode_console_1),
+		cmocka_unit_test(test_get_param_mode_console_2),
+		cmocka_unit_test(test_get_param_mode_console_3),
 	};
 
 	return cmocka_run_group_tests(tests, NULL, NULL);
